+++ crates/compression/src/compressed_block_payload/mod.rs
@@ -1,0 +1,3 @@
+pub mod v0;
+
+pub mod v1;